+++ example/android/app/src/main/kotlin/jonasbark/de/stripepaymentexample/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import io.flutter.app.FlutterActivity
 import io.flutter.plugins.GeneratedPluginRegistrant
 
-class MainActivity : io.flutter.app.FlutterFragmentActivity() {
+class MainActivity : FlutterActivity() {
 
 
     @Override