+++ example/lib/main.dart
@@ -1,5 +1,8 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:stripe_payment/stripe_payment.dart';
+import 'dart:io';
 
 void main() {
   runApp(new MyApp());
@@ -11,82 +14,237 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  var _paymentMethodId = "";
-  var _setupPaymentId = "";
-  var _confirmPaymentId = "";
-  var _confirmNativePay = "";
+  Token _paymentToken;
+  PaymentMethod _paymentMethod;
+  String _error;
+  final String _currentSecret = null; //set this yourself, e.g using curl
+  PaymentIntentResult _paymentIntent;
+  Source _source;
 
-  final _currentSecret = "...";
+  ScrollController _controller = ScrollController();
+
+  final CreditCard testCard = CreditCard(
+    number: '4000002760003184',
+    expMonth: 12,
+    expYear: 21,
+  );
+
+  GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey();
 
   @override
   initState() {
     super.initState();
 
-    StripePayment.setSettings(
-        StripeSettings(publishableKey: "pk_test_yZuUz6Sqm83H4lA7SrlAvYCh003MvJiJlR", merchantIdentifier: "merchant.rbii.stripe-example", androidProductionEnvironment: false));
+    StripePayment.setOptions(
+        StripeOptions(publishableKey: "pk_test_aSaULNS8cJU6Tvo20VAXy6rp", merchantId: "Test", androidPayMode: 'test'));
+  }
+
+  void setError(dynamic error) {
+    _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text(error.toString())));
+    setState(() {
+      _error = error.toString();
+    });
   }
 
   @override
   Widget build(BuildContext context) {
     return new MaterialApp(
       home: new Scaffold(
+        key: _scaffoldKey,
         appBar: new AppBar(
           title: new Text('Plugin example app'),
-        ),
-        body: Column(
-          children: <Widget>[
-            RaisedButton(
-              child: Text("Add Card"),
+          actions: <Widget>[
+            IconButton(
+              icon: Icon(Icons.clear),
               onPressed: () {
-                StripePayment.addSource().then((String token) {
-                  setState(() {
-                    _paymentMethodId = token;
-                  });
+                setState(() {
+                  _source = null;
+                  _paymentIntent = null;
+                  _paymentMethod = null;
+                  _paymentToken = null;
                 });
               },
-            ),
-            Text("Current payment method ID: $_paymentMethodId"),
-            Divider(),
+            )
+          ],
+        ),
+        body: ListView(
+          controller: _controller,
+          padding: const EdgeInsets.all(20),
+          children: <Widget>[
             RaisedButton(
-              child: Text("Setup payment"),
+              child: Text("Create Source"),
               onPressed: () {
-                StripePayment.setupPayment(_paymentMethodId, _currentSecret).then((String token) {
+                StripePayment.createSourceWithParams(SourceParams(
+                  type: 'ideal',
+                  amount: 1099,
+                  currency: 'eur',
+                  returnURL: 'example://stripe-redirect',
+                )).then((source) {
+                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${source.sourceId}')));
                   setState(() {
-                    _setupPaymentId = token;
+                    _source = source;
                   });
-                }).catchError(print);
+                }).catchError(setError);
               },
             ),
-            Text("Current setup payment ID: $_setupPaymentId"),
             Divider(),
             RaisedButton(
-              child: Text("Confirm payment"),
+              child: Text("Create Token with Card Form"),
               onPressed: () {
-                StripePayment.confirmPayment(_paymentMethodId, _currentSecret).then((String token) {
+                StripePayment.paymentRequestWithCardForm(CardFormPaymentRequest()).then((paymentMethod) {
+                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
                   setState(() {
-                    _confirmPaymentId = token;
+                    _paymentMethod = paymentMethod;
                   });
-                }).catchError(print);
+                }).catchError(setError);
               },
             ),
-            Text("Current confirm payment ID: $_confirmPaymentId"),
-            Divider(),
-            LayoutBuilder(
-              builder: (context, constraints) => RaisedButton(
-                child: Text("Native payment"),
-                onPressed: () {
-                  StripePayment.useNativePay(Order(20, 1, 1, "EUR", "Example Store")).then((String token) {
-                    setState(() {
-                      _confirmNativePay = token;
-                    });
-                    StripePayment.confirmNativePay(true);
-                  }).catchError((e) {
-                    Scaffold.of(context).showSnackBar(SnackBar(content: Text(e.toString())));
+            RaisedButton(
+              child: Text("Create Token with Card"),
+              onPressed: () {
+                StripePayment.createTokenWithCard(
+                  testCard,
+                ).then((token) {
+                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${token.tokenId}')));
+                  setState(() {
+                    _paymentToken = token;
                   });
-                },
-              ),
+                }).catchError(setError);
+              },
             ),
-            Text("Native payment: $_confirmNativePay"),
+            Divider(),
+            RaisedButton(
+              child: Text("Create Payment Method with Card"),
+              onPressed: () {
+                StripePayment.createPaymentMethod(
+                  PaymentMethodRequest(
+                    card: testCard,
+                  ),
+                ).then((paymentMethod) {
+                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
+                  setState(() {
+                    _paymentMethod = paymentMethod;
+                  });
+                }).catchError(setError);
+              },
+            ),
+            RaisedButton(
+              child: Text("Create Payment Method with existing token"),
+              onPressed: _paymentToken == null
+                  ? null
+                  : () {
+                      StripePayment.createPaymentMethod(
+                        PaymentMethodRequest(
+                          card: CreditCard(
+                            token: _paymentToken.tokenId,
+                          ),
+                        ),
+                      ).then((paymentMethod) {
+                        _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${paymentMethod.id}')));
+                        setState(() {
+                          _paymentMethod = paymentMethod;
+                        });
+                      }).catchError(setError);
+                    },
+            ),
+            Divider(),
+            RaisedButton(
+              child: Text("Confirm Payment Intent"),
+              onPressed: _paymentMethod == null || _currentSecret == null
+                  ? null
+                  : () {
+                      StripePayment.confirmPaymentIntent(
+                        PaymentIntent(
+                          clientSecret: _currentSecret,
+                          paymentMethodId: _paymentMethod.id,
+                        ),
+                      ).then((paymentIntent) {
+                        _scaffoldKey.currentState
+                            .showSnackBar(SnackBar(content: Text('Received ${paymentIntent.paymentIntentId}')));
+                        setState(() {
+                          _paymentIntent = paymentIntent;
+                        });
+                      }).catchError(setError);
+                    },
+            ),
+            RaisedButton(
+              child: Text("Authenticate Payment Intent"),
+              onPressed: _currentSecret == null
+                  ? null
+                  : () {
+                      StripePayment.authenticatePaymentIntent(clientSecret: _currentSecret).then((paymentIntent) {
+                        _scaffoldKey.currentState
+                            .showSnackBar(SnackBar(content: Text('Received ${paymentIntent.paymentIntentId}')));
+                        setState(() {
+                          _paymentIntent = paymentIntent;
+                        });
+                      }).catchError(setError);
+                    },
+            ),
+            Divider(),
+            RaisedButton(
+              child: Text("Native payment"),
+              onPressed: () {
+                if (Platform.isIOS) {
+                  _controller.jumpTo(450);
+                }
+                StripePayment.paymentRequestWithNativePay(
+                  androidPayOptions: AndroidPayPaymentRequest(
+                    total_price: "1.20",
+                    currency_code: "EUR",
+                  ),
+                  applePayOptions: ApplePayPaymentOptions(
+                    countryCode: 'DE',
+                    currencyCode: 'EUR',
+                    items: [
+                      ApplePayItem(
+                        label: 'Test',
+                        amount: '13',
+                      )
+                    ],
+                  ),
+                ).then((token) {
+                  setState(() {
+                    _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Received ${token.tokenId}')));
+                    _paymentToken = token;
+                  });
+                }).catchError(setError);
+              },
+            ),
+            RaisedButton(
+              child: Text("Complete Native Payment"),
+              onPressed: () {
+                StripePayment.completeNativePayRequest().then((_) {
+                  _scaffoldKey.currentState.showSnackBar(SnackBar(content: Text('Completed successfully')));
+                }).catchError(setError);
+              },
+            ),
+            Divider(),
+            Text('Current source:'),
+            Text(
+              JsonEncoder.withIndent('  ').convert(_source?.toJson() ?? {}),
+              style: TextStyle(fontFamily: "Monospace"),
+            ),
+            Divider(),
+            Text('Current token:'),
+            Text(
+              JsonEncoder.withIndent('  ').convert(_paymentToken?.toJson() ?? {}),
+              style: TextStyle(fontFamily: "Monospace"),
+            ),
+            Divider(),
+            Text('Current payment method:'),
+            Text(
+              JsonEncoder.withIndent('  ').convert(_paymentMethod?.toJson() ?? {}),
+              style: TextStyle(fontFamily: "Monospace"),
+            ),
+            Divider(),
+            Text('Current payment intent:'),
+            Text(
+              JsonEncoder.withIndent('  ').convert(_paymentIntent?.toJson() ?? {}),
+              style: TextStyle(fontFamily: "Monospace"),
+            ),
+            Divider(),
+            Text('Current error: $_error'),
           ],
         ),
       ),