+++ lib/src/source_params.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/material.dart';
+import 'package:stripe_payment/src/token.dart';
+
+class SourceParams {
+  int amount;
+  String currency;
+  String returnURL;
+  String type;
+  String name;
+  String statement_descriptor;
+  String country;
+  String email;
+  CreditCard card;
+
+  SourceParams(
+      {this.amount,
+      this.currency,
+      @required this.returnURL,
+      @required this.type,
+      this.name,
+      this.statement_descriptor,
+      this.country,
+      this.email,
+      this.card});
+
+  factory SourceParams.fromJson(Map<String, dynamic> json) {
+    return SourceParams(
+        amount: json['amount'],
+        currency: json['currency'],
+        returnURL: json['returnURL'],
+        type: json['type'],
+        name: json['name'],
+        statement_descriptor: json['statement_descriptor'],
+        country: json['country'],
+        email: json['email'],
+        card: json['card']);
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    if (this.amount != null) data['amount'] = this.amount;
+    if (this.currency != null) data['currency'] = this.currency;
+    if (this.returnURL != null) data['returnURL'] = this.returnURL;
+    if (this.type != null) data['type'] = this.type;
+    if (this.name != null) data['name'] = this.name;
+    if (this.statement_descriptor != null) data['statement_descriptor'] = this.statement_descriptor;
+    if (this.country != null) data['country'] = this.country;
+    if (this.email != null) data['email'] = this.email;
+    if (this.card != null) data['card'] = this.card.toJson();
+    return data;
+  }
+}